Lots of changes
In short, the core functions are now in separate scripts called by the
just file. This is because the containers will soon be deployed
dinamically as per the users config. This process involves playing
around with lots of enviroment variables and kind of dirty tricks, so
doing everything inside the just file was getting out of hand.
It still looks very ugly, but it works.